requirements.txt
accelerate
einops
wandb
contourpy==1.3.3
cycler==0.12.1
filelock==3.20.2
fonttools==4.61.1
fsspec==2025.12.0
ftfy==6.3.1
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.4.9
matplotlib==3.10.8
more-itertools==10.3.0
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.0
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
pandas==2.3.3
pillow==12.1.0
pyparsing==3.3.1
python-dateutil==2.9.0.post0
pytz==2025.2
regex==2025.11.3
requests==2.32.4
rich==13.9.4
scikit-learn==1.8.0
scipy==1.16.3
seaborn==0.13.2
setuptools==78.1.1
six==1.17.0
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.9.1
torchvision==0.24.1
tqdm==4.67.1
triton==3.5.1
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.3
urllib3==2.5.0
wcwidth==0.2.14
wheel==0.45.1
zstandard==0.23.0
